Split internal readers into a submodules
