+++ pysparkling/sql/internal_utils/readers/__init__.py
@@ -1,0 +1,1 @@
+from pysparkling.sql.internal_utils.readers.common import InternalReader
